+++ target_parquet/helpers.py
@@ -1,0 +1,35 @@
+try:
+    from collections.abc import MutableMapping
+except ImportError:
+    from collections import MutableMapping  # deprecated in Python 3.3
+
+
+def flatten(dictionary, parent_key="", sep="__"):
+    """Function that flattens a nested structure, using the separater given as parameter, or uses '__' as default
+    E.g:
+     dictionary =  {
+                        'key_1': 1,
+                        'key_2': {
+                               'key_3': 2,
+                               'key_4': {
+                                      'key_5': 3,
+                                      'key_6' : ['10', '11']
+                                 }
+                        }
+                       }
+    By calling the function with the dictionary above as parameter, you will get the following strucure:
+        {
+             'key_1': 1,
+             'key_2__key_3': 2,
+             'key_2__key_4__key_5': 3,
+             'key_2__key_4__key_6': "['10', '11']"
+         }
+    """
+    items = []
+    for k, v in dictionary.items():
+        new_key = parent_key + sep + k if parent_key else k
+        if isinstance(v, MutableMapping):
+            items.extend(flatten(v, new_key, sep=sep).items())
+        else:
+            items.append((new_key, str(v) if type(v) is list else v))
+    return dict(items)
